+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.1.1 (09/16/2022)
 
 -   Fixed UI not changing from night to day. Changed bool true to string "true"
+-   Fixed button function error. Changed catchNft to correct function name of catchPkmn
 
 ## v0.1.0 (09/15/2022)